Use full 32bits of hash in flow.
Remove unused internal passing of hash to parse functions.
Leave external prototype alone even though hash is unused.

+++ inetflow.h
@@ -86,7 +86,7 @@ typedef struct _InetFlow {
     guint64 packets;
     InetFlowState state;
     guint family;
-    guint16 hash;
+    guint32 hash;
     guint16 flags;
     guint8 direction;
     guint16 server_port;
